Update to Qt Creator 16.0
Used for "Token_Notification_*" the colors from the terminal.

Applied the same style that "OneDark Pro" theme is using. Namely having
the toolbar colors gray as the "Dark (2024)" theme, and for
"PaletteHighlight*" the "Token_Accent*" values.

Also bundle the Phantom QStyle to replace the Fusion style.

The usage of Phantom QStyle can be turned off by setting
QTCREATOR_USE_DEFAULT_STYLE to 1.